requirements.txt
datasets
sentence-transformers
transformers
torch
faiss-cpu
numpy
pandas
pdfplumber
python-dotenv
tqdm
scikit-learn
 
